requirements.txt
# OpenAI
openai>=1.0.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
elasticsearch>=8.0.0
redis>=4.0.0

# NLP
konlpy>=0.6.0

# Crawling
selenium>=4.0.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# Rate Limiting
slowapi>=0.1.9

# Development
pytest>=7.0.0
black>=23.0.0
ruff>=0.1.0
